[REFAC: #39] - Dividi a CLI em modulos.
close #39
